tools -> gameobjects prepping for restructure
